Move all existing types to types.go
